backend/src/requirements.txt
snowflake-connector-python
snowflake-snowpark-python
pandas
numpy
flask
sqlglot>=25.0.0
sqlfluff>=3.1.0
sqllineage>=1.5.4
dbt-artifacts-parser>=0.6.0
